apps/backend/inss/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.22.3
supabase_auth
# gotrue removido (depreciado)
twilio==8.11.0
reportlab==4.0.9
python-barcode==0.15.1
Pillow==10.2.0
langchain==0.1.6
langchain-openai==0.0.5
# pydantic removido para resolução automática
# pydantic-settings removido para resolução automática
python-dotenv==1.0.0
httpx
pytest==7.4.4
